Developing calendar creation
NEW DEVELOPMENT: In order to allow "confirm and decline" functionality
on for each shop, the shop email must be given.  Therefore,
creating calendars will not suffice.  Shops must be considered as
timekit "resources" (a.k.a. timekit users) in order that the shops
could receive emails of potential customers looking for appointments.
Will develop tomorrow.